Extracted component from modelSelect to seperate files
