+++ src/js/shot-generator-new/components/ModelSelect/ItemSettings.js
@@ -1,0 +1,10 @@
+const GUTTER_SIZE = 5
+const ITEM_WIDTH = 68
+const ITEM_HEIGHT = 132
+
+const IMAGE_WIDTH = ITEM_WIDTH
+const IMAGE_HEIGHT = 100
+
+const NUM_COLS = 4
+
+export {GUTTER_SIZE, ITEM_HEIGHT, ITEM_WIDTH, IMAGE_HEIGHT, IMAGE_WIDTH, NUM_COLS}
